refactor: change root path
