+++ src/routes/playground/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import Playground from '$lib/components/Playground.svelte';
+</script>
+
+<Playground />
